+++ src/main/java/com/realestate/service/ClientService.java
@@ -1,19 +1,19 @@
 package com.realestate.service;
 
+
 import com.realestate.dto.ClientDto;
-import com.realestate.dto.OfferDto;
+import com.realestate.exceptions.CannotDeleteResourceException;
+import com.realestate.exceptions.ResourceNotFoundException;
 import com.realestate.mapper.ClientMapper;
-import com.realestate.mapper.OfferMapper;
 import com.realestate.model.client.Client;
-import com.realestate.model.offer.Offer;
 import com.realestate.repository.ClientRepository;
-import com.realestate.repository.OffersRepository;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import javax.swing.plaf.PanelUI;
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -27,14 +27,19 @@ public class ClientService {
         this.clientMapper = clientMapper;
     }
 
-    public Optional<ClientDto> getClientById(Long id) {
-        return clientRepository.findById(id).map(clientMapper::map);
+    public ClientDto getClientById(Long id) {
+        return clientRepository.findById(id).map(clientMapper::map)
+                .orElseThrow(() -> new ResourceNotFoundException("Client not found"));
     }
 
-    public List<ClientDto> getAllClients() {
-        List<Client> clients = (List<Client>) clientRepository.findAll();
-        List<ClientDto> dtos = clients.stream().map(clientMapper::map).collect(Collectors.toList());
-        return dtos;
+
+    public List<ClientDto> getAllClients(int page, int size) {
+        Pageable pageable = PageRequest.of(page, size);
+        Page<Client> clientPage = clientRepository.findAll(pageable);
+        List<ClientDto> dtos = clientPage.getContent().stream()
+                .map(clientMapper::map)
+                .collect(Collectors.toList());
+        return  dtos;
     }
 
     @Transactional
@@ -45,13 +50,34 @@ public class ClientService {
     }
 
     @Transactional
-    public void updateClient(ClientDto clientDto) {
-        Client client = clientMapper.map(clientDto);
+    public void updateClient(Long id,ClientDto clientDto) {
+        Client client = clientRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Client not found"));
+        updateClient(clientDto, client);
+
         clientRepository.save(client);
     }
 
     public void deleteClient(Long id) {
+        Client client = clientRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Client not found"));
+        if (!client.getOffers().isEmpty()){
+            throw new CannotDeleteResourceException("Cannot delete client with offers");
+        }
         clientRepository.deleteById(id);
+    }
+
+
+    /**
+
+     HELPER METHODS FOR REGISTER
+
+     **/
+
+
+    private static void updateClient(ClientDto clientDto, Client client) {
+        client.setTelephone(clientDto.getTelephone());
+        client.setFirstName(clientDto.getFirstName());
+        client.setLastName(clientDto.getLastName());
+        client.setEmail(clientDto.getEmail());
     }
 
 }
